Nagios: Update reload_nagios logic, remove exec entrypoint
After the move to run the Nagios chart in the hosts pid namespace
or shared PID namespaces, the previous logic for reloading nagios
to update the lists of hosts no longer works when the container is
not run as PID 1. This updates the logic to instead find the
required grandparent PID (Nagios spins up worker processes which
then execute the checks, hence grandparent PID for nagios), and
then issues os.kill on that PID with the SIGHUP signal.

This also removes the exec from the entrypoint to prevent further
complications

Change-Id: I4716ff9df0a00ae6a9865fa59d3a6af090d0a7ed